CORE-8367 Refactor service down/up to IplantWindowInterface
