+++ de-lib/src/main/java/org/iplantc/de/desktop/client/presenter/DesktopWindowManager.java
@@ -5,7 +5,6 @@ import org.iplantc.de.client.models.WindowType;
 import org.iplantc.de.commons.client.ErrorHandler;
 import org.iplantc.de.commons.client.views.window.configs.ConfigFactory;
 import org.iplantc.de.commons.client.views.window.configs.WindowConfig;
-import org.iplantc.de.desktop.client.views.windows.DEAppsWindow;
 import org.iplantc.de.desktop.client.views.windows.IPlantWindowInterface;
 import org.iplantc.de.desktop.client.views.windows.util.WindowFactory;
 import org.iplantc.de.shared.AsyncProviderWrapper;
@@ -176,23 +175,23 @@ public class DesktopWindowManager {
         });
     }
 
-    public void disable(ServiceDown event) {
+    public void serviceDown(ServiceDown event) {
         WindowType type = event.getWindowType();
         for (Widget w : windowManager.getStack()) {
             Window window = (Window) w;
             if (Strings.nullToEmpty(window.getStateId()).startsWith(type.toString())) {
-                DEAppsWindow appsWindow = (DEAppsWindow)window;
-                appsWindow.serviceDown(event.getSelectionHandler());
+                IPlantWindowInterface windowInterface = (IPlantWindowInterface)window;
+                windowInterface.serviceDown(event.getSelectionHandler());
             }
         }
     }
 
-    public void restoreWindow(WindowType windowType) {
+    public void serviceUp(WindowType windowType) {
         for (Widget w : windowManager.getStack()) {
             Window window = (Window) w;
             if (Strings.nullToEmpty(window.getStateId()).startsWith(windowType.toString())) {
-                DEAppsWindow appsWindow = (DEAppsWindow)window;
-                appsWindow.restoreWindow();
+                IPlantWindowInterface windowInterface = (IPlantWindowInterface)window;
+                windowInterface.serviceUp();
             }
         }
     }
