+++ de-lib/src/main/java/org/iplantc/de/desktop/client/views/windows/DEAppsWindow.java
@@ -9,11 +9,9 @@ import org.iplantc.de.commons.client.views.dialogs.AgaveAuthPrompt;
 import org.iplantc.de.commons.client.views.window.configs.AppsWindowConfig;
 import org.iplantc.de.commons.client.views.window.configs.ConfigFactory;
 import org.iplantc.de.commons.client.views.window.configs.WindowConfig;
-import org.iplantc.de.desktop.client.views.widgets.ServiceDownPanel;
 import org.iplantc.de.desktop.shared.DeModule;
 import org.iplantc.de.resources.client.messages.IplantDisplayStrings;
 
-import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
 
 import com.sencha.gxt.widget.core.client.event.SelectEvent;
@@ -26,8 +24,6 @@ public class DEAppsWindow extends IplantWindowBase {
     public static final String APPS = "#apps";
     @Inject UserInfo userInfo;
     private final AppsView.Presenter presenter;
-    private Widget currentWidget;
-    ServiceDownPanel serviceDownPanel;
 
     @Inject
     DEAppsWindow(final AppsView.Presenter presenter, final IplantDisplayStrings displayStrings) {
@@ -84,20 +80,4 @@ public class DEAppsWindow extends IplantWindowBase {
         presenter.setViewDebugId(baseID + AppsModule.Ids.APPS_VIEW);
     }
 
-    public void serviceDown(SelectEvent.SelectHandler handler) {
-        if (currentWidget == null) {
-            currentWidget = getWidget();
-            serviceDownPanel = new ServiceDownPanel();
-            this.setWidget(serviceDownPanel);
-        }
-
-        serviceDownPanel.addHandler(handler);
-    }
-
-    public void restoreWindow() {
-        if (currentWidget != null) {
-            this.setWidget(currentWidget);
-            currentWidget = null;
-        }
-    }
 }
